hireai-main/celery-worker/requirements.txt
FastAPI-SQLAlchemy==0.2.1
SQLAlchemy==2.0.16
openai==0.28.0
python-jose==3.3.0
bcrypt==4.0.1
pyjwt==2.8.0
json-log-formatter==0.5.1
loguru==0.6.0
cryptography==42.0.4
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic-settings==2.2.1
asyncpg==0.28.0
aiofiles==0.6.0
aiohttp==3.9.3
fastapi-oauth2==1.0.0b3
celery==5.3.6
redis==5.0.1
pandas==2.2.0
numpy==1.26.3
opencv-python==4.8.0.76
fpdf2==2.7.7
PyPDF2==3.0.1
halo==0.0.31